% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.r
\name{deg}
\alias{deg}
\title{deg}
\usage{
data(exp)

data(pheno)

final<-filtermatrix(exp)

res<-deg(final,pheno$group,control="NT")
}
\arguments{
\item{group}{group column from phenotype file}

\item{control}{string defining control in pheno$group data}
}
\description{
deg
}
\examples{
 data(exp)
data(pheno)
final<-filtermatrix(exp)
res<-deg(final,pheno$group,control="NT")
head(res)
}
