% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degcofactor.R
\name{degcofactor}
\alias{degcofactor}
\title{degcofactor}
\usage{
data(exp)

data(pheno)

final<-filtermatrix(exp)

res<-degcofactor(final,pheno$group,control="NT",cofactor=pheno$batch)
}
\arguments{
\item{group}{group column from phenotype file}

\item{control}{string defining control in pheno$group data}

\item{cofactor}{string defining cofactor variable in design matrix}
}
\description{
degcofactor
}
\examples{
 data(exp)
data(pheno)
final<-filtermatrix(exp)
res<-degcofactor(final,pheno$group,control="NT",cofactor=pheno$batch)
head(res)
}
