% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtresig.R
\name{filtresig}
\alias{filtresig}
\title{filtresig}
\usage{
data(exp)

data(pheno)

final<-filtermatrix(exp)

res<-deg(final,pheno$group,control="NT")

sig<-filtresig(res)
}
\arguments{
\item{res}{results from limma deg analysis}
}
\description{
filtresig
}
\examples{
 data(exp)
data(pheno)
final<-filtermatrix(exp)
res<-deg(final,pheno$group,control="NT")
sig<-filtresig(res)
dim(sig)
head(sig)
}
