% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcatrans.r
\name{pcatrans}
\alias{pcatrans}
\title{pcatrans}
\usage{
data(exp)

data(pheno)

final<-filtermatrix(exp)

res<-deg(final,pheno$group,control="NT")

sig<-filtresig(res)

process<-reducedf(sig,final,n=20)

pcatrans(process,pheno,group="group",pal="Set1",alpha=0.7,names=TRUE)## palette examples : Set1, Set2, Set3, Pastel1, Pastel2, Dark1, Dark2, Accent
}
\arguments{
\item{group}{group column from phenotype file}

\item{alpha}{tranparency of the points}

\item{names}{see or not names of the samples}

\item{x}{first axis}

\item{y}{second axis}
}
\description{
pcatrans
}
\examples{
data(exp)
data(pheno)
final<-filtermatrix(exp)
variable<-varsig(final,x=1)
res<-deg(variable,pheno$group,control="NT")
sig<-filtresig(res)
process<-reducedf(sig,final,n=20)
pcatrans(process,pheno,group="group",pal="Set1",alpha=0.7,names=TRUE)
pcatrans(variable,pheno,group="group",pal="Set1",alpha=0.7,names=TRUE)
}
