% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reducedf.R
\name{reducedf}
\alias{reducedf}
\title{reducedf}
\usage{
data(exp)

data(pheno)

final<-filtermatrix(exp)

res<-deg(final,pheno$group,control="NT")

process<-reducedf(res,final,n=35)
}
\arguments{
\item{res}{results from limma}

\item{data}{original dataframe output from filtermatrix}

\item{n}{number of genes}
}
\description{
reducedf
}
\examples{
 data(exp)
data(pheno)
final<-filtermatrix(exp)
res<-deg(final,pheno$group,control="NT")
sig<-filtresig(res)
process<-reducedf(sig,final,n=35)
head(process)
}
