% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsig.r
\name{varsig}
\alias{varsig}
\title{varsig}
\usage{
variable<-varsig(final,x=1)
}
\arguments{
\item{x}{number of standard deviation over mean to defined theshold}
}
\description{
varsig
}
\examples{
data(exp)
final<-filtermatrix(exp)
variable<-varsig(final,x=1)
head(variable)
}
