% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vollimma.r
\name{vollimma}
\alias{vollimma}
\title{vollimma}
\usage{
data(exp)

data(pheno)

final<-filtermatrix(exp)

res<-deg(final,pheno$group,control="NT")

vollimma(res,nb=500,fc=0.5,p=0.05,size=4,alpha=1)
}
\arguments{
\item{res}{results from limma deg analysis}

\item{nb}{number of gene names to write}

\item{fc}{fold change threshold}

\item{p}{p-value threshold}

\item{size}{size of the gene labels}

\item{alpha}{transparent points}
}
\description{
vollimma
}
\examples{
data(exp)
data(pheno)
final<-filtermatrix(exp)
res<-deg(final,pheno$group,control="NT")
vollimma(res,nb=500,fc=0.5,p=0.05,size=4,alpha=1)
}
